packages/connections/pyproject.toml
[project]
name = "rhoai-mcp-connections"
version = "0.1.0"
description = "Connections component for RHOAI MCP - Data Connection management"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Adam Miller", email = "admiller@redhat.com" }
]
requires-python = ">=3.10"
keywords = ["mcp", "openshift", "rhoai", "connections", "s3"]

dependencies = [
    "rhoai-mcp-core>=0.1.0,<1.0.0",
]

[project.entry-points."rhoai_mcp.plugins"]
connections = "rhoai_mcp_connections.plugin:create_plugin"

[project.urls]
Homepage = "https://github.com/admiller/rhoai-mcp-prototype"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rhoai_mcp_connections"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint.isort]
known-first-party = ["rhoai_mcp_connections", "rhoai_mcp_core"]
